% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{The application User-Interface}
\usage{
app_ui(request)
}
\arguments{
\item{request}{Internal parameter for `{shiny}`.
DO NOT REMOVE.}
}
\description{
This function creates the user interface for the Shiny application.
It defines the layout, input controls, and output elements that
users will interact with.
}
