% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{create_properties_ui}
\alias{create_properties_ui}
\title{Create UI elements for properties}
\usage{
create_properties_ui(properties, ns, language = "en")
}
\arguments{
\item{properties}{A list of properties}

\item{ns}{A namespace function}

\item{language}{The current language}
}
\value{
A list of UI elements
}
\description{
Create UI elements for properties
}
