% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{create_ui}
\alias{create_ui}
\title{Create UI elements based on parsed schema}
\usage{
create_ui(parsed_schema, ns, language = "en")
}
\arguments{
\item{parsed_schema}{A parsed schema structure}

\item{ns}{A namespace function}

\item{language}{The current language}
}
\value{
A list of UI elements
}
\description{
Create UI elements based on parsed schema
}
