% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{create_widget}
\alias{create_widget}
\title{Create individual widget}
\usage{
create_widget(element, ns = NS(NULL), language = "en")
}
\arguments{
\item{element}{A parsed property structure}

\item{ns}{A namespace function}

\item{language}{The current language}
}
\value{
A UI widget
}
\description{
Create individual widget
}
