% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parser.R
\name{parse_event}
\alias{parse_event}
\title{Parse individual event}
\usage{
parse_event(event, schema)
}
\arguments{
\item{event}{A list representing an event in the schema}
}
\value{
A parsed event structure
}
\description{
Parse individual event
}
