% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parser.R
\name{parse_json_schema}
\alias{parse_json_schema}
\title{Parse JSON schema}
\usage{
parse_json_schema(schema)
}
\arguments{
\item{schema}{A list representing the JSON schema}
}
\value{
A parsed schema structure
}
\description{
Parse JSON schema
}
