% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parser.R
\name{get_multilingual_field}
\alias{get_multilingual_field}
\title{Get multilingual field}
\usage{
get_multilingual_field(obj, field)
}
\arguments{
\item{obj}{A list containing multilingual fields}

\item{field}{The base name of the field}
}
\value{
A list of multilingual values
}
\description{
Get multilingual field
}
