% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{get_select_choices}
\alias{get_select_choices}
\title{Get choices for select inputs}
\usage{
get_select_choices(element, language = "en")
}
\arguments{
\item{element}{The parsed property structure}

\item{language}{The current language}
}
\value{
A named vector of choices
}
\description{
Get choices for select inputs
}
