% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_parser.R
\name{parse_property}
\alias{parse_property}
\title{Parse individual property}
\usage{
parse_property(prop, schema)
}
\arguments{
\item{prop}{A list representing a property in the schema}
}
\value{
A parsed property structure
}
\description{
Parse individual property
}
