% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_ui.R
\name{update_ui_element}
\alias{update_ui_element}
\title{Update UI element}
\usage{
update_ui_element(session, element, value, language = "en")
}
\arguments{
\item{session}{The current Shiny session}

\item{element}{The element to update}

\item{value}{The new value}

\item{language}{The current language}
}
\description{
Update UI element
}
